fix 32-bit build: 0xFFFFFFFF overflows int on 32-bit platforms
Comparisons against math.MaxUint32 (0xFFFFFFFF) using int type fail to
compile on 32-bit platforms (arm/v7, 386) where int is 32-bit. Cast to
int64 for the upper-bound comparison; the check is dead code on 32-bit
anyway since int can never exceed 0x7FFFFFFF there.

+++ filter/u32.go
@@ -19,7 +19,7 @@ func (e *Expr) u32Parse() error {
 	if !ok {
 		return fmt.Errorf("invalid value: %v (expected integer)", e.Val)
 	}
-	if v < 0 || v > 0xFFFFFFFF {
+	if v < 0 || int64(v) > 0xFFFFFFFF {
 		return fmt.Errorf("value out of range: %d", v)
 	}
 	e.Val = uint32(v)
